test: Moved all common files into `tests/lib`
