+++ tests/lib/custom-assertions.js
