+++ tests/lib/error-setup.js
