+++ tests/lib/metrics-tests.js
