+++ tests/lib/promise-resolvers.js
