+++ tests/lib/test-client.js
