+++ tests/versioned/apollo-server-express/errors.test.js
@@ -7,7 +7,7 @@
 
 const test = require('node:test')
 
-const { afterEach, setupExpressTest } = require('../../test-tools')
+const { afterEach, setupExpressTest } = require('../../lib/test-tools')
 
 const errorsTests = require('../errors-tests')
 const { pluginConfig } = errorsTests
