+++ tests/versioned/attributes-tests.js
@@ -7,9 +7,9 @@
 
 const assert = require('node:assert')
 
-const { executeQuery, executeJson } = require('../test-client')
-const { match } = require('../custom-assertions')
-const promiseResolvers = require('../promise-resolvers')
+const { executeQuery, executeJson } = require('../lib/test-client')
+const { match } = require('../lib/custom-assertions')
+const promiseResolvers = require('../lib/promise-resolvers')
 
 const SEGMENT_DESTINATION = 0x20
 const SPAN_DESTINATION = 0x10
