+++ tests/versioned/express-segments-tests.js
@@ -7,10 +7,10 @@
 
 const assert = require('node:assert')
 
-const { executeQuery, executeQueryBatch } = require('../test-client')
+const { executeQuery, executeQueryBatch } = require('../lib/test-client')
 const { checkResult } = require('./common')
-const { assertSegments } = require('../custom-assertions')
-const promiseResolvers = require('../promise-resolvers')
+const { assertSegments } = require('../lib/custom-assertions')
+const promiseResolvers = require('../lib/promise-resolvers')
 
 const ANON_PLACEHOLDER = '<anonymous>'
 const UNKNOWN_OPERATION = '<unknown>'
