+++ tests/versioned/transaction-naming-tests.js
@@ -7,9 +7,9 @@
 
 const assert = require('node:assert')
 
-const { executeQuery, executeQueryBatch } = require('../test-client')
+const { executeQuery, executeQueryBatch } = require('../lib/test-client')
 const { checkResult } = require('./common')
-const promiseResolvers = require('../promise-resolvers')
+const promiseResolvers = require('../lib/promise-resolvers')
 
 const ANON_PLACEHOLDER = '<anonymous>'
